Split the process into steps.
This allows sections to be retried without restarting the whole process.

+++ configure.sh
@@ -13,28 +13,45 @@ set -e
 
 . ./sh/getopts.sh
 
-rm -rf install
-mkdir -p install
+want_step () {
+  [ "$step" = "all" -o "$step" = "$1" ]
+}
 
-. ./sh/downloads.sh
+if want_step download
+then
+  rm -rf install
+  mkdir -p install
 
-echo Running necessary steps as root via sudo...
-ACS_DOMAIN="$baseURL" REALM="$realm" sudo -E /bin/bash ./sh/as-root.sh
+  . ./sh/downloads.sh
+fi
 
-echo "Setting KUBECONFIG for use by k8s tools..."
-export KUBECONFIG="$(realpath ./install/k3s.yaml)"
+if want_step install
+then
+  echo Running necessary steps as root via sudo...
+  ACS_DOMAIN="$baseURL" REALM="$realm" sudo -E /bin/bash ./sh/as-root.sh
 
-. ./sh/kubeseal.sh
+  echo "Setting KUBECONFIG for use by k8s tools..."
+  export KUBECONFIG="$(realpath ./install/k3s.yaml)"
 
-echo Installing Javascript dependencies...
-npm install
+  . ./sh/kubeseal.sh
 
-export KRB5CCNAME=$(mktemp)
-read -p "Please enter the username of a Factory+ administrator user: " KERBUSER
-kinit $KERBUSER
+  echo Installing Javascript dependencies...
+  npm install
+fi
 
-echo Registering edge cluster...
-node src/bootstrap.js "${scheme}://edo.${baseURL}" "$template"
+if want_step join
+then
+  export KRB5CCNAME=$(mktemp)
+  read -p "Please enter the username of a Factory+ administrator user: " KERBUSER
+  kinit $KERBUSER
 
-. ./install/cluster-info.sh
-. ./sh/setup-flux.sh
+  echo Registering edge cluster...
+  node src/bootstrap.js "${scheme}://edo.${baseURL}" "$template"
+fi
+
+if want_step flux
+then
+  export KUBECONFIG="$(realpath ./install/k3s.yaml)"
+  . ./install/cluster-info.sh
+  . ./sh/setup-flux.sh
+fi
